+++ src/main/java/com/zincoid/nullbot/core/model/data/po/StatisticPO.java
@@ -16,7 +16,6 @@ public class StatisticPO {
     private Integer id;
     private Long groupId;
     private Long userId;
-    private String userName;
     private String command;
     private Long visits;
 }