+++ src/main/java/com/zincoid/nullbot/core/service/system/impl/StatisticServiceImpl.java
@@ -1,13 +1,9 @@
 package com.zincoid.nullbot.core.service.system.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
-import com.mikuac.shiro.core.Bot;
-import com.zincoid.nullbot.core.component.tool.BotOperator;
-import com.zincoid.nullbot.core.service.basic.UserService;
 import lombok.RequiredArgsConstructor;
 import com.zincoid.nullbot.core.model.data.po.StatisticDatePO;
 import com.zincoid.nullbot.core.model.data.po.StatisticPO;
-import com.zincoid.nullbot.core.model.data.po.UserPO;
 import com.zincoid.nullbot.core.model.data.vo.StatisticVO;
 import com.zincoid.nullbot.core.mapper.StatisticDateMapper;
 import com.zincoid.nullbot.core.mapper.StatisticMapper;
@@ -26,9 +22,6 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class StatisticServiceImpl implements StatisticService {
 
-    private final BotOperator botOperator;
-    private final UserService userService;
-
     private final StatisticMapper statisticMapper;
     private final StatisticDateMapper statisticDateMapper;
 
@@ -46,15 +39,14 @@ public class StatisticServiceImpl implements StatisticService {
 
     @Override
     @Transactional
-    public void increase(Long groupId, Long userId, String userName, String command) {
+    public void increase(Long groupId, Long userId, String command) {
         StatisticPO statistic = statisticMapper.selectOne(new LambdaQueryWrapper<StatisticPO>()
                 .eq(StatisticPO::getGroupId, groupId)
                 .eq(StatisticPO::getUserId, userId)
                 .eq(StatisticPO::getCommand, command));
         if (statistic == null)
-            statisticMapper.insert(new StatisticPO(null, groupId, userId, userName, command, 1L));
+            statisticMapper.insert(new StatisticPO(null, groupId, userId, command, 1L));
         else {
-            statistic.setUserName(userName);
             statistic.setVisits(statistic.getVisits() + 1);
             statisticMapper.updateById(statistic);
         }
@@ -92,11 +84,9 @@ public class StatisticServiceImpl implements StatisticService {
         while (!currentDate.isAfter(endDate)) {
             // 添加到 X 轴
             xAxis.add(currentDate.format(DateTimeFormatter.ofPattern("MM-dd")));
-
-            // 获取访问量，如果没有数据则为 0
+            // 获取访问量 如果没有数据则为 0
             Long visits = dataMap.getOrDefault(currentDate, 0L);
             data.add(visits);
-
             currentDate = currentDate.plusDays(1);
         }
 
@@ -110,19 +100,12 @@ public class StatisticServiceImpl implements StatisticService {
         }
 
         // 查询访问次数最多的user
-        Bot bot = botOperator.getBot();
         List<Map<String, Object>> topUsers = statisticMapper.selectTopUsers(20);
         List<String> userAxis = new ArrayList<>();
         List<Long> userData = new ArrayList<>();
         for (Map<String, Object> map : topUsers) {
             long userId = Long.parseLong(map.get("user_id").toString());
-            String userName;
-            UserPO user = userService.getById(userId);
-            if (user != null) {
-                userName = user.getName();
-            } else {
-                userName = bot.getStrangerInfo(userId, true).getData().getNickname();
-            }
+            String userName = map.get("user_name").toString();
             userAxis.add(userName + "\n(" + userId + ")");
             userData.add(Long.valueOf(map.get("total_visits").toString()));
         }