+++ src/main/resources/mysql/statistic.sql
@@ -14,7 +14,6 @@ CREATE TABLE `statistic` (
     id INT PRIMARY KEY AUTO_INCREMENT,
     group_id LONG NOT NULL,
     user_id LONG NOT NULL,
-    user_name VARCHAR(50) NOT NULL,
     command VARCHAR(50) NOT NULL,
     visits BIGINT DEFAULT 0
 );